Allow ErrorCode 'ParseFailed'.
The Hexillion documentation states the following:

Parsing happens in several stages, so there may be useful parsed data in the
result even if part of the parsing fails, yielding a ParseFailed error. It’s
important to actually look for the fields you want in the response rather than
simply relying on the value of ErrorCode.

With this change, if the API returns the error code 'ParseFailed', we will
still attempt to parse the XML response.